PostController => update

+++ controllers/PostController.js
@@ -36,9 +36,32 @@ const PostController = {
         }
     },
 
+    async update(req, res) {
+        try {
+            const paramsId = req.params._id;
+            const postById = await Post.findById(paramsId);
+            if (!postById) {
+                return res.status(400).send(`Id ${paramsId} not exists in DB`);
+            };
+            const post = await Post.findByIdAndUpdate(
+                paramsId,
+                req.body,
+                { new: true }
+            );
+          res.status(200).send({message: `Post with id ${paramsId} updated`, post});
+        } catch (error) {
+          console.error(error);
+          res.status(500).send(error);
+        }
+    },
+
     async delete(req, res) {
         try {
             const paramsId = req.params._id;
+            const postById = await Post.findById(paramsId);
+            if (!postById) {
+                return res.status(400).send(`Id ${paramsId} not exists in DB`);
+            };
             const post = await Post.findByIdAndDelete(paramsId);
             res.send({message: `Post with id ${paramsId} deleted`, post});
         } catch (error) {
@@ -46,6 +69,8 @@ const PostController = {
             res.status(500).send({message: `Error trying to remove post with id ${paramsId}`, error});
         }
       }
+
+      // Crear middleware de ID check par ano repetir codigo cada vez qu enos pasen un id concreto por params
 };
 
 module.exports = PostController;
